% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{delete_identities}
\alias{delete_identities}
\title{Deletes identities from an identity pool}
\usage{
delete_identities(IdentityIdsToDelete)
}
\arguments{
\item{IdentityIdsToDelete}{[required] A list of 1-60 identities that you want to delete.}
}
\description{
Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.
}
\details{
You must use AWS Developer credentials to call this API.
}
\section{Accepted Parameters}{
\preformatted{delete_identities(
  IdentityIdsToDelete = list(
    "string"
  )
)
}
}

