% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{describe_identity}
\alias{describe_identity}
\title{Returns metadata related to the given identity, including when the identity was created and any associated linked logins}
\usage{
describe_identity(IdentityId)
}
\arguments{
\item{IdentityId}{[required] A unique identifier in the format REGION:GUID.}
}
\description{
Returns metadata related to the given identity, including when the identity was created and any associated linked logins.
}
\details{
You must use AWS Developer credentials to call this API.
}
\section{Accepted Parameters}{
\preformatted{describe_identity(
  IdentityId = "string"
)
}
}

