% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{describe_identity_pool}
\alias{describe_identity_pool}
\title{Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users}
\usage{
describe_identity_pool(
  IdentityPoolId = "string"
)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}
}
\description{
Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.
}
\details{
You must use AWS Developer credentials to call this API.
}
\examples{

}
