% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{get_id}
\alias{get_id}
\title{Generates (or retrieves) a Cognito ID}
\usage{
get_id(IdentityPoolId, AccountId = NULL, Logins = NULL)
}
\arguments{
\item{IdentityPoolId}{An identity pool ID in the format REGION:GUID.}

\item{AccountId}{A standard AWS account ID (9+ digits).}

\item{Logins}{A set of optional name-value pairs that map provider names to provider tokens. The available provider names for \code{Logins} are as follows:
\itemize{
\item Facebook: \code{graph.facebook.com}
\item Amazon Cognito Identity Provider: \code{cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789}
\item Google: \code{accounts.google.com}
\item Amazon: \code{www.amazon.com}
\item Twitter: \code{api.twitter.com}
\item Digits: \code{www.digits.com}
}}
}
\description{
Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.
}
\details{
This is a public API. You do not need any credentials to call this API.
}
\examples{

}
