% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{get_identity_pool_roles}
\alias{get_identity_pool_roles}
\title{Gets the roles for an identity pool}
\usage{
get_identity_pool_roles(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{An identity pool ID in the format REGION:GUID.}
}
\description{
Gets the roles for an identity pool.
}
\details{
You must use AWS Developer credentials to call this API.
}
\examples{

}
