% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{get_open_id_token}
\alias{get_open_id_token}
\title{Gets an OpenID token, using a known Cognito ID}
\usage{
get_open_id_token(IdentityId, Logins = NULL)
}
\arguments{
\item{IdentityId}{[required] A unique identifier in the format REGION:GUID.}

\item{Logins}{A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access\_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito Identity Provider, or any other OpenId Connect provider, always include the \code{id_token}.}
}
\description{
Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.
}
\details{
The OpenId token is valid for 15 minutes.

This is a public API. You do not need any credentials to call this API.
}
\section{Accepted Parameters}{
\preformatted{get_open_id_token(
  IdentityId = "string",
  Logins = list(
    "string"
  )
)
}
}

