% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{list_identities}
\alias{list_identities}
\title{Lists the identities in a pool}
\usage{
list_identities(
  IdentityPoolId = "string",
  MaxResults = 123,
  NextToken = "string",
  HideDisabled = TRUE|FALSE
)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}

\item{MaxResults}{[required] The maximum number of identities to return.}

\item{NextToken}{A pagination token.}

\item{HideDisabled}{An optional boolean parameter that allows you to hide disabled identities. If omitted, the ListIdentities API will include disabled identities in the response.}
}
\description{
Lists the identities in a pool.
}
\details{
You must use AWS Developer credentials to call this API.
}
\examples{

}
