% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_package.R
\docType{package}
\name{paws.cognitoidentity-package}
\alias{paws.cognitoidentity}
\alias{paws.cognitoidentity-package}
\title{paws.cognitoidentity: Amazon Cognito Identity}
\description{
Amazon Cognito is a web service that delivers
    scoped temporary credentials to mobile devices and other untrusted
    environments. Amazon Cognito uniquely identifies a device and supplies
    the user with a consistent identity over the lifetime of an
    application.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
