% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{unlink_identity}
\alias{unlink_identity}
\title{Unlinks a federated identity from an existing account}
\usage{
unlink_identity(IdentityId, Logins, LoginsToRemove)
}
\arguments{
\item{IdentityId}{A unique identifier in the format REGION:GUID.}

\item{Logins}{A set of optional name-value pairs that map provider names to provider tokens.}

\item{LoginsToRemove}{Provider names to unlink from this identity.}
}
\description{
Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.
}
\details{
This is a public API. You do not need any credentials to call this API.
}
\examples{

}
