% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitoidentity_operations.R
\name{update_identity_pool}
\alias{update_identity_pool}
\title{Updates a user pool}
\usage{
update_identity_pool(IdentityPoolId, IdentityPoolName,
  AllowUnauthenticatedIdentities, SupportedLoginProviders = NULL,
  DeveloperProviderName = NULL, OpenIdConnectProviderARNs = NULL,
  CognitoIdentityProviders = NULL, SamlProviderARNs = NULL)
}
\arguments{
\item{IdentityPoolId}{An identity pool ID in the format REGION:GUID.}

\item{IdentityPoolName}{A string that you provide.}

\item{AllowUnauthenticatedIdentities}{TRUE if the identity pool supports unauthenticated logins.}

\item{SupportedLoginProviders}{Optional key:value pairs mapping provider names to provider app IDs.}

\item{DeveloperProviderName}{The "domain" by which Cognito will refer to your users.}

\item{OpenIdConnectProviderARNs}{A list of OpendID Connect provider ARNs.}

\item{CognitoIdentityProviders}{A list representing an Amazon Cognito Identity User Pool and its client ID.}

\item{SamlProviderARNs}{An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.}
}
\description{
Updates a user pool.
}
\details{
You must use AWS Developer credentials to call this API.
}
\examples{

}
